rename: fix file names
